Fix vote type component
